feat: add client pages
