+++ fe/app/(authenticated)/appointments/page.tsx
@@ -1,62 +1,10 @@
-"use client";
+import { requireStaffRole } from "@/lib/auth-guard";
 
-import RoleGuard from "@/components/auth/RoleGuard";
-import InfiniteScroll from "@/components/InfiniteScroll";
-import AppointmentForm from "@/components/new/forms/AppointmentForm";
-import { useInfiniteAppointments } from "@/hooks/appointments/useAppointments";
-import { flattenPages } from "@/lib/utils";
-import { Appointment } from "@/types/appointment";
+import AppointmentPageClient from "./page.client";
 
-function AppointmentPageContent() {
-  const {
-    data,
-    isLoading,
-    isError,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-  } = useInfiniteAppointments();
+export default async function AppointmentPage() {
+  // Protect this page - only allow staff roles
+  await requireStaffRole();
 
-  const appointments = flattenPages<Appointment>(data?.pages || []);
-  console.log({
-    data,
-    appointments,
-  });
-  return (
-    <div className="p-6">
-      <h1 className="mb-4 text-2xl font-semibold">Appointments</h1>
-
-      <AppointmentForm mode="create" onSuccess={() => {}} />
-
-      <hr className="my-6" />
-
-      {isLoading ? (
-        <p>Loading...</p>
-      ) : isError ? (
-        <p>Something went wrong.</p>
-      ) : (
-        <InfiniteScroll
-          fetchNextPage={fetchNextPage}
-          hasNextPage={hasNextPage}
-          isFetchingNextPage={isFetchingNextPage}
-        >
-          <ul className="space-y-3">
-            {appointments.map((appt, index) => (
-              <li key={`${appt.id}-${index}`} className="rounded border p-3">
-                {appt.appointmentStatus}
-              </li>
-            ))}
-          </ul>
-        </InfiniteScroll>
-      )}
-    </div>
-  );
-}
-
-export default function AppointmentPage() {
-  return (
-    <RoleGuard allowedRoles={["Admin", "Doctor", "Staff"]}>
-      <AppointmentPageContent />
-    </RoleGuard>
-  );
+  return <AppointmentPageClient />;
 }
