+++ fe/app/(authenticated)/medicine-inventory-entries/page.tsx
@@ -1,54 +1,10 @@
-"use client";
+import { requireStaffRole } from "@/lib/auth-guard";
 
-import RoleGuard from "@/components/auth/RoleGuard";
-import InfiniteScroll from "@/components/InfiniteScroll";
-import { useInfiniteMedicineInventoryEntries } from "@/hooks/medicine-inventory-entries/useMedicineInventoryEntries";
-import { flattenPages } from "@/lib/utils";
+import MedicineInventoryEntryPageClient from "./page.client";
 
-function MedicineInventoryEntryPageContent() {
-  const {
-    data,
-    isLoading,
-    isError,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-  } = useInfiniteMedicineInventoryEntries();
+export default async function MedicineInventoryEntryPage() {
+  // Protect this page - only allow staff roles
+  await requireStaffRole();
 
-  const entries = flattenPages(data?.pages || []);
-
-  return (
-    <div className="p-6">
-      <h1 className="mb-4 text-2xl font-semibold">
-        Medicine Inventory Entries
-      </h1>
-      {isLoading ? (
-        <p>Loading...</p>
-      ) : isError ? (
-        <p>Something went wrong.</p>
-      ) : (
-        <InfiniteScroll
-          fetchNextPage={fetchNextPage}
-          hasNextPage={hasNextPage}
-          isFetchingNextPage={isFetchingNextPage}
-        >
-          <ul className="space-y-3">
-            {entries.map((item, index) => (
-              <li key={index} className="rounded border p-3">
-                {item.companyName}
-              </li>
-            ))}
-          </ul>
-        </InfiniteScroll>
-      )}
-    </div>
-  );
-}
-
-export default function MedicineInventoryEntryPage() {
-  return (
-    <RoleGuard allowedRoles={["Admin", "Doctor", "Staff"]}>
-      <MedicineInventoryEntryPageContent />
-    </RoleGuard>
-  );
+  return <MedicineInventoryEntryPageClient />;
 }
